Ignore poisoned mutex on server for rendering
The mutex in question merely ensures that only one Blender instance is allowed to render at a time. It does not guard any data, so it's safe to ignore whether it is poisoned or not. Now the server can keep processing render requests after a Blender instance has crashed (in my case caused by running out of VRAM).