Accept render request only when render slot is free
The client now first sends a general render request, which is accepted by the server once it's free to render the next frame. Only then the client tells the server which particular frame to render.

This change makes the scheduling of render requests more responsive to the available rendering resources and should therefore lead to better decisions on which servers to send requests to:
If one were to render two frames and have two servers available for rendering, prior to this commit, each server would be sent one render request for one frame each from the start. If one of the servers turns out to be busy with other requests, it may end up taking longer for it to deal with those first and then rendering and sending back our frame than just letting the other server deal with both frames. The latter should be the observed behavior now.

The change is also somewhat of a prerequisite for implementing the forwarding of render requests to peer servers, as the forwarding server can now accept additional render requests on behalf of its peers once they are free to render more.