Use single render thread instead of one Blender instance for each client connection
Now only one Blender instance is used throughout the program, both for querying the local system information and for all render requests. This restructuring should make it easier to implement other types of renderers in the future, like peer servers and standalone BRPy instances that run on other devices but which can access the same resources via a network share.